+++ app/build.gradle
@@ -61,6 +61,7 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
+
     implementation "com.google.dagger:hilt-android:$daggerHilt_version"
     kapt "com.google.dagger:hilt-compiler:$daggerHilt_version"
