+++ app/src/main/java/com/rodcollab/cliq/collections/bookings/domain/OnQueryTextChangeUseCase.kt
@@ -1,4 +1,4 @@
-package com.rodcollab.cliq.collections.bookings.form
+package com.rodcollab.cliq.collections.bookings.domain
 
 import com.rodcollab.cliq.collections.clients.model.ClientItem
